Vertical slicing
se crea modulo usuarios y se pasan sus capas a este
dentro del paquete usuarios tambien hay infraestructura, aplicacion y dominio
