--- src/main/java/com/asisge/consultifybackend/aplicacion/dto/Dto.java
@@ -1,6 +1,0 @@
-package com.asisge.consultifybackend.aplicacion.dto;
-
-public interface Dto {
-
-    boolean validarDto();
-}
