+++ src/main/java/com/asisge/consultifybackend/auth/aplicacion/manejador/ManejadorServicioAutenticacion.java
@@ -4,8 +4,8 @@ import com.asisge.consultifybackend.auth.aplicacion.dto.AuthenticationRequest;
 import com.asisge.consultifybackend.auth.aplicacion.dto.AuthenticationResponse;
 import com.asisge.consultifybackend.auth.aplicacion.servicio.ServicioAutenticacion;
 import com.asisge.consultifybackend.auth.dominio.puerto.RepositorioAutorizacion;
-import com.asisge.consultifybackend.dominio.modelo.Usuario;
-import com.asisge.consultifybackend.dominio.modelo.UsuarioAutenticado;
+import com.asisge.consultifybackend.usuarios.dominio.modelo.Usuario;
+import com.asisge.consultifybackend.usuarios.dominio.modelo.UsuarioAutenticado;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
