+++ src/main/java/com/asisge/consultifybackend/auth/aplicacion/manejador/ServicioUserDetailsImpl.java
@@ -1,7 +1,7 @@
 package com.asisge.consultifybackend.auth.aplicacion.manejador;
 
-import com.asisge.consultifybackend.dominio.modelo.UsuarioAutenticado;
-import com.asisge.consultifybackend.dominio.puerto.RepositorioUsuario;
+import com.asisge.consultifybackend.usuarios.dominio.modelo.UsuarioAutenticado;
+import com.asisge.consultifybackend.usuarios.dominio.puerto.RepositorioUsuario;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
