+++ src/main/java/com/asisge/consultifybackend/auth/dominio/puerto/RepositorioAutorizacion.java
@@ -1,6 +1,6 @@
 package com.asisge.consultifybackend.auth.dominio.puerto;
 
-import com.asisge.consultifybackend.dominio.modelo.UsuarioAutenticado;
+import com.asisge.consultifybackend.usuarios.dominio.modelo.UsuarioAutenticado;
 
 public interface RepositorioAutorizacion {
     UsuarioAutenticado buscarPorCorreo(String correo);
