+++ src/main/java/com/asisge/consultifybackend/auth/infraestructura/controlador/ControladorAutorizacion.java
@@ -3,7 +3,7 @@ package com.asisge.consultifybackend.auth.infraestructura.controlador;
 import com.asisge.consultifybackend.auth.aplicacion.dto.AuthenticationRequest;
 import com.asisge.consultifybackend.auth.aplicacion.dto.AuthenticationResponse;
 import com.asisge.consultifybackend.auth.aplicacion.manejador.ManejadorServicioAutenticacion;
-import com.asisge.consultifybackend.dominio.modelo.Usuario;
+import com.asisge.consultifybackend.usuarios.dominio.modelo.Usuario;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
