--- src/main/java/com/asisge/consultifybackend/dominio/modelo/Rol.java
@@ -1,8 +1,0 @@
-package com.asisge.consultifybackend.dominio.modelo;
-
-public enum Rol {
-    ADMIN,
-    ASESOR,
-    CLIENTE
-
-}
