+++ src/main/java/com/asisge/consultifybackend/usuarios/aplicacion/dto/CambioEstadoDto.java
@@ -1,0 +1,21 @@
+package com.asisge.consultifybackend.usuarios.aplicacion.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class CambioEstadoDto implements Dto {
+
+    private Long idUsuario;
+    private String identificacion;
+    private Boolean activo;
+
+    @Override
+    public boolean validarDto() {
+        return this.identificacion != null &&
+                this.activo != null;
+    }
+}
