+++ src/main/java/com/asisge/consultifybackend/usuarios/aplicacion/dto/Dto.java
@@ -1,0 +1,6 @@
+package com.asisge.consultifybackend.usuarios.aplicacion.dto;
+
+public interface Dto {
+
+    boolean validarDto();
+}
