+++ src/main/java/com/asisge/consultifybackend/usuarios/dominio/modelo/Rol.java
@@ -1,0 +1,8 @@
+package com.asisge.consultifybackend.usuarios.dominio.modelo;
+
+public enum Rol {
+    ADMIN,
+    ASESOR,
+    CLIENTE
+
+}
