+++ src/main/java/com/asisge/consultifybackend/usuarios/dominio/puerto/RepositorioUsuario.java
@@ -1,0 +1,33 @@
+package com.asisge.consultifybackend.usuarios.dominio.puerto;
+
+import com.asisge.consultifybackend.usuarios.dominio.modelo.Usuario;
+import com.asisge.consultifybackend.usuarios.dominio.modelo.UsuarioAutenticado;
+
+import java.util.List;
+
+public interface RepositorioUsuario {
+
+    List<Usuario> buscarTodos();
+
+    List<UsuarioAutenticado> buscarTodosUsuariosAutenticados();
+
+    Usuario buscarUsuarioPorId(Long idUsuario);
+
+    UsuarioAutenticado buscarUsuarioPorIdentificacion(String identificacion);
+
+    UsuarioAutenticado buscarUsuarioPorCorreo(String correo);
+
+    UsuarioAutenticado buscarPorCorreoOUsername(String correoOUsername);
+
+    UsuarioAutenticado crearUsuarioAutenticado(UsuarioAutenticado usuarioAutenticado);
+
+    void eliminarUsuario(String identificacion);
+
+    void cambiarContrasena(UsuarioAutenticado usuarioAutenticado);
+
+    UsuarioAutenticado editarInformacionBasica(Usuario aGuardar);
+
+    UsuarioAutenticado editarCorreo(Usuario existente);
+
+    UsuarioAutenticado cambiarEstado(UsuarioAutenticado usuario, boolean activo);
+}
